refactor(parser): extract type taxonomy into types/supported
Plan 2, step 3 of N. Move the type taxonomy — `SupportedType`,
`UnsupportedType`, `TypeResolutionError`, and their `Display`
impls — into `crates/toolr-core/src/parser/types/supported.rs`
(125 lines).

`types/mod.rs` shrinks 1059 → 944 lines. The `serde::{Deserialize,
Serialize}` import dropped from mod.rs (only `SupportedType` used
it, and that lives in supported.rs now).

Public surface preserved via `pub use supported::{SupportedType,
TypeResolutionError, UnsupportedType};` in `types/mod.rs`.

Tests: 24 / 24 parser::types::tests still pass; cargo check
--workspace clean.
